added 3 dummy pages for psych profiles
