+++ frontend/src/pages/index.tsx
@@ -1,12 +1,12 @@
 import Head from 'next/head'
 import { Inter } from 'next/font/google'
-import PsychiatristProfile from "../pages/pro_profile"
+import PsychiatristProfile from "./pro_profile1"
 
 export default function Home() {
   return (
     <div className="mx-auto">
       <Head>
-        <PsychiatristProfile /> 
+        <PsychiatristProfile />
       </Head>
     </div>
   );
